ASSERT in RenderBox::instrinsicScrollbarLogicalWidth opening the inspector https://bugs.webkit.org/show_bug.cgi?id=136929
Reviewed by Simon Fraser.

This is a regression from http://trac.webkit.org/changeset/173668
This code is wrong after that change because it assumes that overflow:scroll will
always have a scrollbar. That is no longer the case. We need to check to make sure
it’s an overflow:scroll that does NOT have overflow:auto behavior.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::instrinsicScrollbarLogicalWidth):



Identifier: 154748@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173734 268f45cc-cd09-0410-ab3c-d52691b4dbfc