WebContent crash in WebKit::WebPage::expandedRangeFromHandle https://bugs.webkit.org/show_bug.cgi?id=136969
Reviewed by David Kilzer.

The crash was caused by m_currentBlockSelection and frame.selection().selection().toNormalizedRange()
both being null and the subsequent lines of code assuming currentRange to be not null.

Replace the assertion by an early exit since the range could be null here.

In the future, we should figure out if there are other things we need to do in addition
to checking the nullity when this race condition is hit.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::computeExpandAndShrinkThresholdsForHandle):


Identifier: 154784@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173788 268f45cc-cd09-0410-ab3c-d52691b4dbfc