ASSERTION FAILED: columnCount in WebCore::CSSParser::parseGridTemplateAreasRow https://bugs.webkit.org/show_bug.cgi?id=136945
Reviewed by Sergio Villar Senin.

Source/WebCore:

Checking out whether the grid-template-areas value contains a white-space only
string, which is not valid as it does not produce a cell token.

Test: fast/css-grid-layout/grid-template-areas-empty-string-crash.html

* css/CSSParser.cpp:
(WebCore::CSSParser::parseGridTemplateAreasRow):

LayoutTests:

Testing the different types of white-space only strings for the grid-template-areas
property, which is not valid as it does not produce a cell token.

* fast/css-grid-layout/grid-template-areas-empty-string-crash-expected.txt: Added.
* fast/css-grid-layout/grid-template-areas-empty-string-crash.html: Added.


Identifier: 154920@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173965 268f45cc-cd09-0410-ab3c-d52691b4dbfc