Add the baseline implementation of :not(selectorList) https://bugs.webkit.org/show_bug.cgi?id=137548
Reviewed by Andreas Kling.

Source/WebCore:

This patch extend the pseudo class :not() to support the new definition
in CSS Selectors level 4.

Instead of supporting a single simple selector, :not() now support
an arbitrary selector list, excepted pseudo elements.

One side effect of this change is :visited and :link are no longer supported
inside :not(). The reason has to do with complexity and performance.
Since :not() and :matches() can have arbitrary selector list, keeping track of
:visited and :link can have many combination superposing, some valid, some invalid.

Supporting :visited and :link should be possible by chaning the way we handle
them for style resolution, but given the limited use cases for such features
I'll wait to see if there is any interest before changing everything.

This patch only covers SelectorChecker to keep things simple. The CSS JIT
will come later.

Tests: fast/css/not-basics.html
       fast/css/parsing-css-not-1.html
       fast/css/parsing-css-not-2.html
       fast/css/parsing-css-not-3.html
       fast/css/parsing-css-not-4.html
       fast/css/parsing-css-not-5.html
       fast/css/parsing-css-not-6.html
       fast/css/parsing-css-not-7.html
       fast/css/parsing-css-not-8.html
       fast/css/parsing-css-not-9.html
       fast/css/parsing-css-nth-child-of-4.html
       fast/selectors/not-basics.html
       fast/selectors/not-nested.html
       fast/selectors/not-selector-list.html

* css/CSSGrammar.y.in:
Rename selectorListDoesNotMatchAnyPseudoElement() to selectorListDoesNotMatchAnyPseudoElement()
since both :not() and :nth-child(An+B of selectorList) have the same requirements at the moment.

* css/CSSParserValues.cpp:
(WebCore::selectorListMatchesPseudoElement):
The patch adding :matches() introduced the possibility of putting pseudo element
arbitrary deep in a sequence of nested :matches(). Since this case is forbidded
by :not() and :nth-child(An+B of selectorList), selectorListMatchesPseudoElement()
now becomes recursive to find those pseudo elements.

* css/CSSSelector.cpp:
(WebCore::CSSSelector::selectorText):
CSSOM for the extended :not().

* css/SelectorChecker.cpp:
(WebCore::SelectorChecker::checkOne):
(WebCore::SelectorChecker::determineLinkMatchType):
* cssjit/SelectorCompiler.cpp:

(WebCore::SelectorCompiler::addPseudoClassType):
Disable the new capabilities from the JIT for now.

LayoutTests:

There are 3 big kind of changes to the tests:
1) The new implementation intentionally breaks :not() of :visited and :link.
   Some test for the basic :not(:link) and :not(:visited) are consequently failing.
2) Some tests were ensuring that :not() only support simple selectors, those are
   becoming irrelevant.
3) New test for the feature.

* TestExpectations:
This patch intentionally breaks :not(:link). The test fast/history/link-inside-not.html
time out since it is waiting for the link to match.

* fast/css/invalid-not-with-simple-selector-sequence-expected.txt: Removed.
* fast/css/invalid-not-with-simple-selector-sequence.html: Removed.
This test originally came from IEtestcenter. None of it is useful, all the "invalid"
cases are now fully functional selectors.

* fast/css/css-selector-text-expected.txt:
* fast/css/css-selector-text.html:
* fast/css/css-set-selector-text-expected.txt:
* fast/css/css-set-selector-text.html:
Parsing and CSS OM tests for the new :not().

* fast/css/parsing-css-matches-5-expected.txt:
* fast/css/parsing-css-matches-5.html:
* fast/css/parsing-css-matches-6-expected.txt:
* fast/css/parsing-css-matches-6.html:
* fast/css/parsing-css-matches-7-expected.txt:
* fast/css/parsing-css-matches-7.html:
* fast/css/parsing-css-matches-8-expected.txt:
* fast/css/parsing-css-matches-8.html:
Update the tests of :matches() to account for the extended :not().

* fast/dom/SelectorAPI/resig-SelectorsAPI-test-expected.txt:
Some cases covered by this test are now valid. The test is kept as is for
its historical value, but all the FAIL cover valid use cases in Level 4.

* platform/mac/css3/selectors3/html/css3-modsel-61-expected.txt:
* platform/mac/css3/selectors3/html/css3-modsel-83-expected.txt:
* platform/mac/css3/selectors3/xhtml/css3-modsel-61-expected.txt:
* platform/mac/css3/selectors3/xhtml/css3-modsel-83-expected.txt:
* platform/mac/css3/selectors3/xml/css3-modsel-61-expected.txt:
* platform/mac/css3/selectors3/xml/css3-modsel-83-expected.txt:
* platform/mac/fast/selectors/061-expected.txt:
* platform/mac/fast/selectors/083-expected.txt:
Those tests are changing because of the breakage of :link and :visited.

* fast/css/not-basics-expected.html: Added.
* fast/css/not-basics.html: Added.
Basic use cases of the extended :not() that should always style correctly.

* fast/css/parsing-css-not-1-expected.txt: Added.
* fast/css/parsing-css-not-1.html: Added.
* fast/css/parsing-css-not-2-expected.txt: Added.
* fast/css/parsing-css-not-2.html: Added.
* fast/css/parsing-css-not-3-expected.txt: Added.
* fast/css/parsing-css-not-3.html: Added.
* fast/css/parsing-css-not-4-expected.txt: Added.
* fast/css/parsing-css-not-4.html: Added.
* fast/css/parsing-css-not-5-expected.txt: Added.
* fast/css/parsing-css-not-5.html: Copied from LayoutTests/fast/css/parsing-css-matches-5.html.
* fast/css/parsing-css-not-6-expected.txt: Added.
* fast/css/parsing-css-not-6.html: Copied from LayoutTests/fast/css/parsing-css-matches-6.html.
* fast/css/parsing-css-not-7-expected.txt: Copied from LayoutTests/fast/css/parsing-css-matches-7-expected.txt.
* fast/css/parsing-css-not-7.html: Copied from LayoutTests/fast/css/parsing-css-matches-7.html.
* fast/css/parsing-css-not-8-expected.txt: Copied from LayoutTests/fast/css/parsing-css-matches-8-expected.txt.
* fast/css/parsing-css-not-8.html: Copied from LayoutTests/fast/css/parsing-css-matches-8.html.
* fast/css/parsing-css-not-9-expected.txt: Added.
* fast/css/parsing-css-not-9.html: Copied from LayoutTests/fast/css/parsing-css-matches-8.html.
Parsing of :not(). A lot of them were inspired by Yusuke Suzuki's test suite
for :matches(). A lot of test case apply to :not(), the main difference is that
pseudo elements must fail inside :not().

* fast/css/parsing-css-nth-child-of-4-expected.txt: Added.
* fast/css/parsing-css-nth-child-of-4.html: Added.
With the introduction of :matches(), it is possible to have arbitrarily deep pseudo elements.
The tests for :not() cover the case fixed with this patch, this addition covers :nth-child(of).

* fast/selectors/not-basics-expected.txt: Added.
* fast/selectors/not-basics.html: Added.
* fast/selectors/not-nested-expected.txt: Added.
* fast/selectors/not-nested.html: Added.
* fast/selectors/not-selector-list-expected.txt: Added.
* fast/selectors/not-selector-list.html: Added.
Test coverage for the new features.


Identifier: 155417@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174535 268f45cc-cd09-0410-ab3c-d52691b4dbfc